+++ packages/jugg/CHANGELOG.md
@@ -5,6 +5,8 @@
 - [ ] refactor: 集中式管理 webpack 插件、loader 等等的配置项，并对外提供获取、修改的途径
 - feat: 新增 `jugg inspect` 命令用于查看配置
 - feat: 命令行支持 `mode=` 以配置 `NODE_ENV`，默认 `development`
+  - 但当运行 `dev`、`build` 时，`NODE_ENV` 总是为 `development` 和 `production`
+- feat: 除了自动载入 `.env` 环境变量文件，还会根据 `mode` 载入 `.mode.env` 文件
 
 ## 0.2.0 - 2019-11-11