+++ packages/jugg/CHANGELOG.md
@@ -1,8 +1,8 @@
 # CHANGELOG
 
-## 0.2.1 -
+## 0.2.1-alpha.0 - 2020-01-02
 
-- [ ] refactor: 集中式管理 webpack 插件、loader 等等的配置项，并对外提供获取、修改的途径
+- refactor: 集中式管理 webpack 插件、loader 等等的配置项，并对外提供获取、修改的途径
 - feat: 新增 `jugg inspect` 命令用于查看配置
 - feat: 命令行支持 `mode=` 以配置 `NODE_ENV`，默认 `development`
   - 但当运行 `dev`、`build` 时，`NODE_ENV` 总是为 `development` 和 `production`